--- src/main/java/com/example/project02/controller/testController.java
@@ -1,15 +1,0 @@
-package com.example.project02.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api")
-public class testController {
-
-    @GetMapping
-    public String test(){
-        return "test";
-    }
-}
