+++ src/main/java/com/example/project02/dto/Request.java
@@ -1,4 +1,4 @@
-package com.example.project02.model;
+package com.example.project02.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
