+++ src/main/java/com/example/project02/dto/SmsCertification.java
@@ -1,4 +1,4 @@
-package com.example.project02.entity.dto;
+package com.example.project02.dto;
 
 import lombok.AccessLevel;
 import lombok.Builder;
