+++ src/main/java/com/example/project02/service/sms/SmsCertificationService.java
@@ -1,8 +1,7 @@
 package com.example.project02.service.sms;
 
-import com.example.project02.entity.dto.SmsCertification;
 import com.example.project02.exception.AuthenticationNumberMismatchException;
-import com.example.project02.model.SmsCertificationDao;
+import com.example.project02.dao.SmsCertification;
 import com.example.project02.properties.AppProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +21,7 @@ import static com.example.project02.service.sms.coolSmsConstants.SMS_TYPE;
 @RequiredArgsConstructor
 public class SmsCertificationService {
 
-    private final SmsCertificationDao smsCertificationDao;
+    private final SmsCertification smsCertification;
     private final AppProperties appProperties;
 
     private DefaultMessageService messageService;
@@ -61,20 +60,20 @@ public class SmsCertificationService {
         SingleMessageSendingRequest sendingRequest = new SingleMessageSendingRequest(message);
         SingleMessageSentResponse response = messageService.sendOne(sendingRequest);
 
-        smsCertificationDao.createSmsCertification(phone, randomNumber);
+        smsCertification.createSmsCertification(phone, randomNumber);
     }
 
 
-    public void verifySms(SmsCertification requestDto) {
+    public void verifySms(com.example.project02.dto.SmsCertification requestDto) {
         if (isVerify(requestDto)) {
             throw new AuthenticationNumberMismatchException("인증번호가 일치하지 않습니다.");
         }
-        smsCertificationDao.removeSmsCertification(requestDto.getPhone());
+        smsCertification.removeSmsCertification(requestDto.getPhone());
     }
 
-    public boolean isVerify(SmsCertification requestDto) {
-        return !(smsCertificationDao.hasKey(requestDto.getPhone()) &&
-            smsCertificationDao.getSmsCertification(requestDto.getPhone())
+    public boolean isVerify(com.example.project02.dto.SmsCertification requestDto) {
+        return !(smsCertification.hasKey(requestDto.getPhone()) &&
+            smsCertification.getSmsCertification(requestDto.getPhone())
                 .equals(requestDto.getCertificationNumber()));
     }
 }
